% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gofchisq.R
\name{gofchisq}
\alias{gofchisq}
\title{Goodness of fit test}
\usage{
gofchisq(x, p, conf.level = 0.95)
}
\arguments{
\item{x}{a vector of observed}

\item{p}{probability of each group}

\item{conf.level}{confidence level}
}
\value{
output for goodness of fit test
}
\description{
This function is the goodness of fit test
}
\examples{
x=c(12,9,10,7,12)
prob=c(1/5,1/5,1/5,1/5,1/5)  #1:1:1:1:1
gofchisq(x=x,p=prob)

}
\references{
Chernoff, H.; Lehmann, E. L.(1954) <doi:10.1214/aoms/1177728726>.
}
