% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indchisq.R
\name{indchisq}
\alias{indchisq}
\title{Independence test}
\usage{
indchisq(O, conf.level = 0.95)
}
\arguments{
\item{O}{an observed matrix has a rows and b columns}

\item{conf.level}{confidence level}
}
\value{
output for independence test
}
\description{
This function is for independence test
}
\examples{
v <- c(80,60,150,50,40,20)
X<- matrix(v,ncol=2,byrow = TRUE)  # 3x2
indchisq(X)

}
\references{
Plackett, R. L. (1983). <doi:10.2307/1402731>.
}
