% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotchisq.R
\name{plotchisq}
\alias{plotchisq}
\title{Plot of Chi-squared distribution}
\usage{
plotchisq(df = 8)
}
\arguments{
\item{df}{degrees of freedom}
}
\value{
The figure of Chi-squared distribution with k degrees of freedom
}
\description{
The plot of Chi-squared distribution with k degrees of freedom
}
\examples{
plotchisq(df=10)

}
